requirements.txt
streamlit>=1.32.0
joblib>=1.3.2
numpy>=1.24.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0 
flask==2.3.3
flask-cors==4.0.0
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.24.3
pymongo==4.5.0
python-dotenv==1.0.0
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
datasets==2.15.0 
requests